refactor(starknet_integration_tests): move node component configs to seperate mod
commit-id:505b2210
