+++ crates/starknet_integration_tests/src/lib.rs
@@ -2,6 +2,7 @@ pub mod flow_test_setup;
 pub mod integration_test_setup;
 pub mod integration_test_utils;
 pub mod monitoring_utils;
+pub mod node_component_configs;
 pub mod sequencer_manager;
 pub mod sequencer_simulator_utils;
 pub mod state_reader;
